+++ src/renderer/containers/AppBar.tsx
@@ -1,22 +1,19 @@
-import { CSSProperties } from 'react';
 import * as React from 'react';
 import { connect, Dispatch } from 'react-redux';
-import { KeycloakProfile } from 'keycloak-js';
 import {
     Button,
     Navbar,
     NavbarDivider,
     NavbarGroup,
     Popover,
-    PopoverPosition, Tooltip,
+    PopoverPosition,
 } from '@blueprintjs/core';
 
 import * as actions from '../actions';
 import { State } from '../state';
 import cateIcon from '../resources/cate-icon-128.png';
-import UserMenu from './UserMenu';
 import WorkspacesMenu from './WorkspacesMenu';
-import FilesMenu from './FilesMenu';
+// import FilesMenu from './FilesMenu';
 import HelpMenu from './HelpMenu';
 
 
@@ -27,33 +24,31 @@ interface IDispatch {
 }
 
 interface IAppBarProps {
-    userProfile: KeycloakProfile | null,
-    pwaInstallPromotionVisible: boolean;
 }
 
 // noinspection JSUnusedLocalSymbols
 function mapStateToProps(state: State): IAppBarProps {
     return {
-        userProfile: state.communication.userProfile,
-        pwaInstallPromotionVisible: state.control.pwaInstallPromotionVisible,
     };
 }
 
 
 const _AppBar: React.FC<IAppBarProps & IDispatch> = (
     {
-        userProfile,
-        pwaInstallPromotionVisible,
         dispatch,
     }
 ) => {
 
-    const handlePreferencesClick = () => {
-        dispatch(actions.showPreferencesDialog());
+    const handleOpenInBrowser = () => {
+        window.open(window.location.href, "_blank");
     };
 
-    const handleShowPwaInstallPrompt = () => {
-        dispatch(actions.showPwaInstallPrompt() as any);
+    const handleShutdownServer = () => {
+        dispatch(actions.showShutdownDialog());
+    };
+
+    const handlePreferencesClick = () => {
+        dispatch(actions.showPreferencesDialog());
     };
 
     return (
@@ -63,39 +58,28 @@ const _AppBar: React.FC<IAppBarProps & IDispatch> = (
                 <h2 style={TITLE_STYLE}>Cate - ESA CCI Toolbox</h2>
             </NavbarGroup>
             <NavbarGroup align="right">
-                {pwaInstallPromotionVisible && (
-                    <React.Fragment>
-                        <Button
-                            className="bp3-minimal"
-                            icon="desktop"
-                            intent="success"
-                            onClick={handleShowPwaInstallPrompt}
-                        >
-                            Install App
-                        </Button>
-                        <NavbarDivider/>
-                    </React.Fragment>
-                )}
                 <Popover content={<WorkspacesMenu/>} position={PopoverPosition.BOTTOM}>
                     <Button className="bp3-minimal" rightIcon={'caret-down'}>Workspaces</Button>
                 </Popover>
-                <Popover content={<FilesMenu/>} position={PopoverPosition.BOTTOM}>
-                    <Button className="bp3-minimal" rightIcon={'caret-down'}>Files</Button>
-                </Popover>
+                {/*<Popover content={<FilesMenu/>} position={PopoverPosition.BOTTOM}>*/}
+                {/*    <Button className="bp3-minimal" rightIcon={'caret-down'}>Files</Button>*/}
+                {/*</Popover>*/}
                 <Popover content={<HelpMenu/>} position={PopoverPosition.BOTTOM}>
                     <Button className="bp3-minimal" rightIcon={'caret-down'}>Help</Button>
                 </Popover>
                 <NavbarDivider/>
-                {userProfile !== null && (
-                    <Popover content={<UserMenu/>} position={PopoverPosition.BOTTOM}>
-                        <Tooltip content={<UserInfo userProfile={userProfile}/>}>
-                            <Button
-                                className="bp3-minimal"
-                                icon={<Avatar userProfile={userProfile}/>}
-                            />
-                        </Tooltip>
-                    </Popover>
-                )}
+                {isRunningInIFrame() &&
+                    <Button
+                        className="bp3-minimal"
+                        icon='share'
+                        onClick={handleOpenInBrowser}
+                    />
+                }
+                <Button
+                    className="bp3-minimal"
+                    icon='offline'
+                    onClick={handleShutdownServer}
+                />
                 <Button
                     className="bp3-minimal"
                     icon='cog'
@@ -109,54 +93,12 @@ const _AppBar: React.FC<IAppBarProps & IDispatch> = (
 const AppBar = connect(mapStateToProps)(_AppBar);
 export default AppBar;
 
-const AVATAR_STYLE: CSSProperties = {
-    width: 28,
-    height: 28,
-    flex: 'none',
-    borderRadius: 14,
-    color: 'white',
-    display: 'flex',
-    textAlign: 'center',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#29A634',
-};
 
-interface AvatarProps {
-    userProfile: KeycloakProfile;
-}
-
-function Avatar({userProfile}: AvatarProps) {
-    const name = userProfile.firstName || userProfile.username || userProfile.email;
-    const letter = ((name && name.length && name[0]) || 'U').toUpperCase();
-    return (<div style={AVATAR_STYLE}>{letter}</div>);
-}
-
-
-interface UserInfoProps {
-    userProfile: KeycloakProfile;
-}
-
-function UserInfo({userProfile}: UserInfoProps) {
-    const text: React.ReactNode[] = [];
-    if (userProfile.username) {
-        text.push(<b>{userProfile.username}</b>);
+// https://stackoverflow.com/questions/326069/how-to-identify-if-a-webpage-is-being-loaded-inside-an-iframe-or-directly-into-t
+function isRunningInIFrame(): boolean {
+    try {
+        return window.self !== window.top;
+    } catch (e) {
+        return true;
     }
-    if (userProfile.firstName) {
-        if (userProfile.lastName) {
-            text.push(`${userProfile.firstName} ${userProfile.lastName}`);
-        } else {
-            text.push(userProfile.firstName);
-        }
-    } else if (userProfile.lastName) {
-        text.push(userProfile.lastName);
-    }
-    if (userProfile.email) {
-        if (userProfile.emailVerified) {
-            text.push(userProfile.email);
-        } else {
-            text.push(`${userProfile.email} (not verified)`);
-        }
-    }
-    return (<div>{text.map((t, i) => <div key={i}>{t}</div>)}</div>);
 }