+++ src/renderer/containers/AppMainPage.tsx
@@ -5,12 +5,13 @@ import { useMatomo } from '@datapunt/matomo-tracker-react'
 
 import GdprBanner from './GdprBanner';
 import { isElectron } from '../electron';
-import { FileSystemAPI } from '../webapi/apis';
+import { FileSystemAPI } from '../webapi';
 import AppBar from './AppBar';
 import ChooseWorkspaceDialog, { DELETE_WORKSPACE_DIALOG_ID, OPEN_WORKSPACE_DIALOG_ID } from './ChooseWorkspaceDialog';
 import GlobeView from './GlobeView'
 import FigureView from './FigureView';
-import ServiceAutoCloseDialog from './ServiceAutoCloseDialog';
+import ServiceAutoShutdownDialog from './ServiceAutoShutdownDialog';
+import ServiceShutdownDialog from './ServiceShutdownDialog';
 import TableView from './TableView';
 import DataSourcesPanel from './DataSourcesPanel';
 import OperationsPanel from './OperationsPanel';
@@ -97,7 +98,7 @@ function mapStateToPropsApplication(state: State): IApplicationPageProps {
     return {
         forceAppBar: state.session.forceAppBar,
         fileSystemAPI: selectors.fileSystemAPISelector(state),
-        appServiceMode: state.communication.userProfile ? 'cloud' : 'local',
+        appServiceMode: window.location.pathname.indexOf("/proxy/") > 0 ? 'cloud' : 'local',
         appServiceURL: state.communication.webAPIServiceURL,
         appServiceHostOS: state.communication.webAPIServiceInfo && state.communication.webAPIServiceInfo.hostOS,
         appServiceVersion: state.communication.webAPIServiceInfo && state.communication.webAPIServiceInfo.version,
@@ -170,7 +171,8 @@ const _AppMainPage: React.FC<IApplicationPageProps & IDispatch> = (
             <FileUploadDialog/>
             <FileDownloadDialog/>
             <SaveWorkspaceAsDialog/>
-            <ServiceAutoCloseDialog/>
+            <ServiceShutdownDialog/>
+            <ServiceAutoShutdownDialog/>
             <ChooseWorkspaceDialog dialogId={OPEN_WORKSPACE_DIALOG_ID}/>
             <ChooseWorkspaceDialog dialogId={DELETE_WORKSPACE_DIALOG_ID}/>
             <OperationStepDialog id={NEW_CTX_OPERATION_STEP_DIALOG_ID}/>