+++ src/renderer/main.tsx
@@ -4,23 +4,31 @@ import { Provider as StoreProvider } from 'react-redux';
 import { applyMiddleware, createStore, Middleware, Store } from 'redux';
 import { createLogger } from 'redux-logger';
 import thunkMiddleware from 'redux-thunk';
-import Keycloak, { KeycloakInitOptions, KeycloakConfig } from 'keycloak-js'
-import { KeycloakProvider } from '@react-keycloak/web'
 
 import * as actions from './actions'
-import AppRouter from './containers/AppRouter';
 import { stateReducer } from './reducers';
 import { State } from './state';
-import { isElectron } from './electron';
-import { CONFIG } from '../config';
+import AppMainPageWrapper from './containers/AppMainPageWrapper';
+import { connectWebAPIService } from './actions';
+import { DEFAULT_SERVICE_URL } from './initial-state';
 
 
-const keycloak = Keycloak(getKeycloakConfig());
-
-const keycloakProviderInitConfig: KeycloakInitOptions = {
-    onLoad: 'check-sso',
-    enableLogging: true,
-};
+function getServiceUrl() {
+    let serviceUrl: string = DEFAULT_SERVICE_URL;
+    const search = new URLSearchParams(window.location.search);
+    if (search.has('serviceUrl')) {
+        serviceUrl = search.get('serviceUrl');
+    } else {
+        const origin = window.location.origin;
+        const path = window.location.pathname;
+        if (path.endsWith('/app')) {
+            serviceUrl = origin + path.substring(0, path.length - 4);
+        } else if (path.endsWith('/app/')) {
+            serviceUrl = origin + path.substring(0, path.length - 5);
+        }
+    }
+    return serviceUrl;
+}
 
 export function main() {
     const middlewares: Middleware[] = [thunkMiddleware];
@@ -45,58 +53,14 @@ export function main() {
 
     const middleware = applyMiddleware(...middlewares);
     const store = createStore(stateReducer, middleware) as Store<State>;
-
-    const onKeycloakEvent = (event, error) => {
-        console.debug('onKeycloakEvent', event, error);
-    }
-
-    const onKeycloakTokens = (tokens) => {
-        console.debug('onKeycloakTokens', tokens);
-    }
-
-    // Fetch hub status from GitHub
-    const deployment = CONFIG.webApi.endpointUrl.includes('stage') || CONFIG.webApi.endpointUrl.includes('dev') ? 'development' : 'production';
-    fetch(`https://raw.githubusercontent.com/CCI-Tools/cate-status/main/${deployment}.json`,
-          {mode: 'cors'})
-        .then(response =>
-                  response.json())
-        .then(hubStatus =>
-                  store.dispatch(actions.updateHubStatus(
-                      {...hubStatus, deployment})))
-        .catch(e => console.error(e));
+    store.dispatch(connectWebAPIService(getServiceUrl()) as any);
 
     ReactDOM.render(
         (
-            <KeycloakProvider
-                keycloak={keycloak}
-                initConfig={keycloakProviderInitConfig}
-                onEvent={onKeycloakEvent}
-                onTokens={onKeycloakTokens}
-            >
-                <StoreProvider store={store}>
-                    <AppRouter/>
-                </StoreProvider>
-            </KeycloakProvider>
+          <StoreProvider store={store}>
+              <AppMainPageWrapper/>
+          </StoreProvider>
         ),
         document.getElementById('root')
     );
-
-    if (!isElectron()) {
-        //
-        // Desktop-PWA app install, see https://web.dev/customize-install/
-        //
-        window.addEventListener('beforeinstallprompt', (event: Event) => {
-            // Update UI notify the user they can install the PWA
-            store.dispatch(actions.showPwaInstallPromotion(event));
-            console.log('BEFORE INSTALL PROMPT:', event);
-        });
-    }
-}
-
-function getKeycloakConfig(): KeycloakConfig {
-    const {url, realm, clientId} = CONFIG.auth;
-    if (!realm || !url || !clientId) {
-        throw new Error('Missing or incomplete KeyCloak configuration');
-    }
-    return {realm, url, clientId};
 }