+++ src/renderer/webapi/apis/ServiceShutdownAPI.ts
@@ -1,0 +1,7 @@
+
+export class ServiceShutdownAPI {
+    shutdown(serviceURL: string): Promise<null> {
+        return fetch(serviceURL + "/exit")
+            .then(() => null);
+    }
+}